src/cloud/requirements.txt
# API Framework
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0.post2

# Database
psycopg2-binary==2.9.6
asyncpg==0.27.0
sqlalchemy==2.0.15
alembic==1.11.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Vector Database
qdrant-client==1.1.1

# Cryptography
pyNaCl==1.5.0
cryptography==40.0.2

# Processing
numpy==1.24.3
pandas==2.0.2
ray==2.5.1

# Embeddings
sentence-transformers==2.2.2
torch==2.0.1

# NLP/LLM
langchain==0.0.190

# Monitoring
prometheus-client==0.17.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-prometheus==1.18.0

# Utilities
pyyaml==6.0
tenacity==8.2.2
orjson==3.9.1
